pyproject.toml
[project]
name = "mas-for-spanish-language-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pydantic-ai>=0.2.10",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
